Move `textNodes` into DOM test utilities for reuse
